requirements.txt
flask
paho-mqtt
urllib3